reduce clang-15 warnings (#4840)
* reduce clang-15 warnings

* update bslmf_invokeresult_cpp03.00.t.cpp

* correct warnings

+++ groups/bsl/bslalg/bslalg_arrayprimitives.t.cpp
@@ -630,7 +630,7 @@ char getValue(const Uint64& u64)
 
 void setValue(float *pf, char ch)
 {
-    *pf = static_cast<int>(ch);
+    *pf = static_cast<short>(ch);
 }
 
 void setValue(double *pf, char ch)
