+++ groups/bsl/bslstl/bslstl_chrono.t.cpp
@@ -192,7 +192,8 @@ int main(int argc, char *argv[])
 
 // Finally, stream the two objects to 'stdout':
 
-        printf("one day is %lld hours\n", hours_in_a_day.count());
+        printf("one day is %lld hours\n",
+               static_cast<long long>(hours_in_a_day.count()));
         printf("half an hour is %.1f hours\n",
                static_cast<double>(halfhour.count()));
 #else
